requirements.txt
dj-database-url==0.5.0
Django==2.2.17
django-geojson==3.1.0
django-leaflet==0.27.1
django-stdimage==5.1.1
Pillow==8.1.0
progressbar2==3.53.1
psycopg2==2.8.6
python-decouple==3.4
python-utils==2.4.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
